로그인한 유저만 기능을 사용할 수 있도록 수정

+++ server/clothes/views.py
@@ -35,9 +35,11 @@ class ClothesView(APIView):
     )
     def get(self, request):
         try:
-            queryset = Clothes.objects.all()
-            serializer = ClothesSerializer(queryset, many=True)
-            return Response(serializer.data, status=status.HTTP_200_OK)
+            if request.user.is_authenticated:
+                queryset = Clothes.objects.all()
+                serializer = ClothesSerializer(queryset, many=True)
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
 
@@ -75,18 +77,20 @@ class ClothesView(APIView):
     )
     def post(self, request):
         try:
-            context = {
-                'category': request.data['category'],
-                'cloth_type': request.data['cloth_type'],
-                'season': request.data['season'],
-                'gender': request.data['gender'],
-                'image': dec_b64_img(request.data)
-            }
-            serializer = ClothesSerializer(data=context)
-            if serializer.is_valid():
-                queryset = serializer.save()
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            if request.user.is_authenticated:
+                context = {
+                    'category': request.data['category'],
+                    'cloth_type': request.data['cloth_type'],
+                    'season': request.data['season'],
+                    'gender': request.data['gender'],
+                    'image': dec_b64_img(request.data)
+                }
+                serializer = ClothesSerializer(data=context)
+                if serializer.is_valid():
+                    queryset = serializer.save()
+                    return Response(serializer.data, status=status.HTTP_201_CREATED)
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
 
@@ -116,9 +120,11 @@ class ClothesDetailView(APIView):
     )
     def get(self, request, pk):
         try:
-            queryset = Clothes.objects.get(pk=pk)
-            serializer = ClothesSerializer(queryset)
-            return Response(serializer.data, status=status.HTTP_200_OK)
+            if request.user.is_authenticated:
+                queryset = Clothes.objects.get(pk=pk)
+                serializer = ClothesSerializer(queryset)
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
 
@@ -162,12 +168,14 @@ class ClothesDetailView(APIView):
     )
     def put(self, request, pk):
         try:
-            queryset = Clothes.objects.get(pk=pk)
-            serializer = ClothesSerializer(queryset, data=request.data)
-            if serializer.is_valid():
-                serializer.save()
-                return Response(serializer.data, status=status.HTTP_202_ACCEPTED)
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            if request.user.is_authenticated:
+                queryset = Clothes.objects.get(pk=pk)
+                serializer = ClothesSerializer(queryset, data=request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    return Response(serializer.data, status=status.HTTP_202_ACCEPTED)
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
 
@@ -190,9 +198,11 @@ class ClothesDetailView(APIView):
     )
     def delete(self, request, pk):
         try:
-            queryset = Clothes.objects.get(pk=pk)
-            queryset.delete()
-            return Response({'message': 'Cloth deleted'}, status=status.HTTP_202_ACCEPTED)
+            if request.user.is_authenticated:
+                queryset = Clothes.objects.get(pk=pk)
+                queryset.delete()
+                return Response({'message': 'Cloth deleted'}, status=status.HTTP_202_ACCEPTED)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
 
@@ -242,39 +252,41 @@ class ClothesRecommendationView(APIView):
     )
     def get(self, request, gender):
         try:
-            gender_str = ["female", "male"]
-            weather_data = open_weather_api()[1]
+            if request.user.is_authenticated:
+                gender_str = ["female", "male"]
+                weather_data = open_weather_api()[1]
 
-            if weather_data["temperature"] >= 30:
-                season = "summer"
-            elif weather_data["temperature"] >= 20:
-                season = "autumn"
-            elif weather_data["temperature"] >= 10:
-                season = "spring"
-            else:
-                season = "winter"
+                if weather_data["temperature"] >= 30:
+                    season = "summer"
+                elif weather_data["temperature"] >= 20:
+                    season = "autumn"
+                elif weather_data["temperature"] >= 10:
+                    season = "spring"
+                else:
+                    season = "winter"
 
-            queryset = Clothes.objects.exclude(gender=gender_str[1-int(gender)])
-            queryset = queryset.filter(season=season)
+                queryset = Clothes.objects.exclude(gender=gender_str[1-int(gender)])
+                queryset = queryset.filter(season=season)
 
-            cloth_top = queryset.filter(category="top").order_by("?").first()
-            cloth_bottom = queryset.filter(category="bottom").order_by("?").first()
-            cloth_coat = queryset.filter(category="coat").order_by("?").first()
+                cloth_top = queryset.filter(category="top").order_by("?").first()
+                cloth_bottom = queryset.filter(category="bottom").order_by("?").first()
+                cloth_coat = queryset.filter(category="coat").order_by("?").first()
 
-            serialized_top = ClothesSerializer(cloth_top).data
-            serialized_bottom = ClothesSerializer(cloth_bottom).data
-            serialized_coat = ClothesSerializer(cloth_coat).data
+                serialized_top = ClothesSerializer(cloth_top).data
+                serialized_bottom = ClothesSerializer(cloth_bottom).data
+                serialized_coat = ClothesSerializer(cloth_coat).data
 
-            context = {
-                "top": modifying_image_path(serialized_top),
-                "bot": modifying_image_path(serialized_bottom),
-                "coat": modifying_image_path(serialized_coat),
-                "gender": gender_str[gender],
-                "w_data": weather_data
-            }
+                context = {
+                    "top": modifying_image_path(serialized_top),
+                    "bot": modifying_image_path(serialized_bottom),
+                    "coat": modifying_image_path(serialized_coat),
+                    "gender": gender_str[gender],
+                    "w_data": weather_data
+                }
 
-            serializer = ClorareSerializer(context)
-            return Response(serializer.data, status=status.HTTP_200_OK)
+                serializer = ClorareSerializer(context)
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
 
@@ -305,11 +317,13 @@ class ClothesCleanUpView(APIView):
     )
     def post(self, request):
         try:
-            if request.data['key'] == "delete all clothes data":
-                queryset = Clothes.objects.all()
-                queryset.delete()
-                return Response({'message': 'Cloth deleted'}, status=status.HTTP_202_ACCEPTED)
-            return Response({'message': 'Incorrect delete key'}, status=status.HTTP_403_FORBIDDEN)
+            if request.user.is_authenticated:
+                if request.data['key'] == "delete all clothes data":
+                    queryset = Clothes.objects.all()
+                    queryset.delete()
+                    return Response({'message': 'Cloth deleted'}, status=status.HTTP_202_ACCEPTED)
+                return Response({'message': 'Incorrect delete key'}, status=status.HTTP_403_FORBIDDEN)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
